No need of banks anymore in eCommerce.
No need of bank transaction, in ebanking when each transaction occurs between the user,mediater(paypal),bank and all of them should be connected. But in this system once we buy points from mediater(paypal) we could buy any product using the points instead of accessing the bank(No need of bank transaction anymore). The mediater will just reduce only our points from our Point-Balance. And no more taxes, once during buying of points it is enough to be paid the tax but not in buying or selling. This will be more useful for the High end trading which delt with hugh amount.